:truck: Move Android files to correct packages and rename plugin channels
Android files are now all in package social.openbook.app

Plugins now all use okuna.io as channel prefix
