+++ android/app/src/main/java/social/openbook/app/util/InputStreamSupplier.java
@@ -1,4 +1,4 @@
-package com.example.openbook.util;
+package social.openbook.app.util;
 
 import java.io.FileNotFoundException;
 import java.io.InputStream;
